momentum-bot-discord-v3/requirements.txt
alpaca-trade-api>=3.2.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
apscheduler>=3.10.4
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.10.0
